added tests for fit and fit_transform on input types
